+++ CHANGELOG.md
@@ -38,6 +38,7 @@ In addition to original Keep-a-Changelog, we use following rules:
 - impl `FromStr` for `Record` and `DataSection` https://github.com/ricosjp/ruststep/pull/140
 
 ### Changed
+- Use `Record` struct in `Parameter::Typed` https://github.com/ricosjp/ruststep/pull/192
 - Generate Holder struct for TYPE declaration with simple type. https://github.com/ricosjp/ruststep/pull/186
 - Replace the methods `xxx_iter` of `Tables` with `xxx_holder`.  https://github.com/ricosjp/ruststep/pull/187
 - Implement `AsRef` and `AsMut` for `XXAny`. https://github.com/ricosjp/ruststep/pull/180